requirements.txt
opencv-python~=4.10
numpy~=2.0
PyYAML~=6.0
pillow~=10.3
requests~=2.31
scipy~=1.13
utils~=1.0